requirements.txt
apkutils==2.0.4
brotli==1.2.0
brotlicffi==1.2.0.0
compiler==0.2.0
ConfigParser==7.2.0
cryptography==46.0.3
Cython==3.2.3
dl==0.1.0
docutils==0.22.4
easyocr==1.7.2
h2==4.2.0
HTMLParser==0.0.2
hypothesis==6.148.8
importlib_metadata==8.7.0
importlib_metadata==8.0.0
joblib==1.5.1
js==1.0
keyring==25.7.0
lockfile==0.12.2
lxml==5.4.0
matplotlib==3.10.8
mock==5.2.0
mtrand==0.1
mypy==1.19.1
numarray==1.5.1
Numeric==24.2
opencv_python==4.12.0.88
ordereddict==1.1
pickle5==0.0.12
Pillow==12.0.0
protobuf==6.33.2
psutil==7.0.0
PyGetWindow==0.0.9
PyInstaller==6.17.0
pyodide==0.0.2
pyOpenSSL==25.3.0
pytesseract==0.3.13
pytest==9.0.2
pytz==2025.2
pywin32==310
PyYAML==6.0.2
PyYAML==6.0.3
scikit_learn==1.8.0
simplejson==3.20.1
thread==2.0.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.15.0
ultralytics==8.3.163
xmlrpclib==1.0.1
zipp==3.19.2
zipp==3.23.0
zstandard==0.24.0
